+++ app/src/main/java/com/ringov/yamblzweather/model/internet/data/Main.java
@@ -8,7 +8,7 @@ public class Main {
 
     @SerializedName("temp")
     @Expose
-    private Double temp;
+    private Float temp;
     @SerializedName("pressure")
     @Expose
     private Integer pressure;
@@ -22,11 +22,11 @@ public class Main {
     @Expose
     private Double tempMax;
 
-    public Double getTemp() {
+    public Float getTemp() {
         return temp;
     }
 
-    public void setTemp(Double temp) {
+    public void setTemp(Float temp) {
         this.temp = temp;
     }
